+++ src/main/java/racingcar/commander/MovementCommander.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.commander;
+
+import racingcar.domain.MoveCommand;
 
 @FunctionalInterface
 public interface MovementCommander {
