+++ src/main/java/racingcar/domain/Cars.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
+import racingcar.commander.MovementCommander;
 
 public class Cars {
