+++ src/main/java/racingcar/domain/MoveCommand.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public enum MoveCommand {
     MOVE {
